Adiçao do Scanner nas Classes

+++ src/ProjetosUNIT/DesafiosAula01/Atividade04/Salario.java
@@ -9,6 +9,6 @@ public class Salario {
 
     public double calcularSalariosMinimos(){
         double salariosMinimos = salario / 1320;
-        return (int) salariosMinimos;
+        return salariosMinimos;
     }
 }
